requirements.txt
Pillow==9.3.0
pyyaml==6.0
tqdm==4.65.0
scikit-learn==1.2.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cu118
torchvision==0.15.2+cu118
pandas==2.0.1
tensorboardX==2.6